requirements.txt
absl-py==0.2.2
astor==0.6.2
attrs==17.4.0
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
enum34==1.1.6
gast==0.2.0
gpflow==1.1.0
grpcio==1.12.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jupyter-client==5.2.2
jupyter-core==4.4.0
markdown==2.6.11
matplotlib==2.1.2
multipledispatch==0.4.9
networkx==2.1
numpy==1.14.5
opencv-python==3.4.1.15
pandas==0.22.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pillow==5.0.0
pluggy==0.6.0
progressbar2==3.37.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.5.2
py==1.5.2
pyarrow==0.9.0
pygments==2.2.0
pyparsing==2.2.0
pytest==3.4.1
python-dateutil==2.6.1
python-utils==2.3.0
pytz==2018.3
pywavelets==0.5.2
pyzmq==17.0.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tabulate==0.8.2
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
tensorpack==0.8.6
termcolor==1.1.0
tornado==4.5.3
tqdm==4.23.4
traitlets==4.3.2
wcwidth==0.1.7
werkzeug==0.14.1
xlrd==1.2.0
